Node Join API and CLI
```
kadalu cluster create mycluster
kadalu node join server1 http://localhost:3000 -c mycluster
kadalu node join server2 http://localhost:3000 -c mycluster
kadalu node join server3 http://localhost:3000 -c mycluster
```
