+++ types/src/moana_types.cr
@@ -22,6 +22,24 @@ module MoanaTypes
     end
   end
 
+  struct NodeRequest
+    include JSON::Serializable
+
+    property name = "", endpoint = "", cluster_name = ""
+
+    def initialize
+    end
+  end
+
+  struct Node
+    include JSON::Serializable
+
+    property id = "", name = "", endpoint = "", addresses = [] of String, token = ""
+
+    def initialize
+    end
+  end
+
   struct Error
     include JSON::Serializable
